+++ src/e2e/tatum.spec.ts
@@ -8,7 +8,7 @@ describe('Tatum Init', () => {
     it('Testnet', async () => {
       const tatum = await TatumSDK.init({
         network: Network.Testnet,
-        rpc: { ignoreLoadBalancing: true },
+        rpc: { ignoreLoadBalancing: true, useStaticUrls: true },
       })
       await e2eUtil.subscriptions.testAddressBasedSubscription(tatum, Chain.Bitcoin, TestConst.ADDRESSES.TESTNET[Chain.Bitcoin], tatum.notification.subscribe.addressEvent)
     })
@@ -16,7 +16,7 @@ describe('Tatum Init', () => {
     it('Mainnet', async () => {
       const tatum = await TatumSDK.init({
         network: Network.Mainnet,
-        rpc: { ignoreLoadBalancing: true },
+        rpc: { ignoreLoadBalancing: true, useStaticUrls: true },
       })
       await e2eUtil.subscriptions.testAddressBasedSubscription(tatum, Chain.Bitcoin, TestConst.ADDRESSES.MAINNET[Chain.Bitcoin], tatum.notification.subscribe.addressEvent)
     })
@@ -31,11 +31,11 @@ describe('Tatum Init', () => {
     it('IP auth', async () => {
       const mainnet = await TatumSDK.init({
         network: Network.Mainnet,
-        rpc: { ignoreLoadBalancing: true },
+        rpc: { ignoreLoadBalancing: true, useStaticUrls: true },
       })
       const testnet = await TatumSDK.init({
         network: Network.Testnet,
-        rpc: { ignoreLoadBalancing: true },
+        rpc: { ignoreLoadBalancing: true, useStaticUrls: true },
       })
       await e2eUtil.subscriptions.testAddressBasedSubscription(mainnet, Chain.Bitcoin, TestConst.ADDRESSES.MAINNET[Chain.Bitcoin], mainnet.notification.subscribe.addressEvent)
       await e2eUtil.subscriptions.testAddressBasedSubscription(testnet, Chain.Bitcoin, TestConst.ADDRESSES.TESTNET[Chain.Bitcoin], testnet.notification.subscribe.addressEvent)